Bookapp: adding the screens and trying to see the problems in the code

+++ app/src/main/java/com/example/bookapp/data/local/dao/BookDao.kt
@@ -19,7 +19,7 @@ interface BookDao {
     fun getAllBooks(): Flow<List<BookEntity>>
 
     @Query("SELECT * FROM books WHERE id = :bookId")
-    suspend fun getBookById(bookId: Int): BookEntity?
+    suspend fun getBookById(bookId: Long): BookEntity?
 
     @Query("SELECT * FROM books WHERE isFavorite = 1 ORDER BY createdAt DESC")
     fun getFavoriteBooks(): Flow<List<BookEntity>>
@@ -31,17 +31,17 @@ interface BookDao {
     suspend fun updateBook(book: BookEntity)
 
     @Query("UPDATE books SET isFavorite = :isFavorite WHERE id = :bookId")
-    suspend fun updateFavoriteStatus(bookId: Int, isFavorite: Boolean)
+    suspend fun updateFavoriteStatus(bookId: Long, isFavorite: Boolean)
 
     @Query("UPDATE books SET rating = :rating, review = :review WHERE id = :bookId")
-    suspend fun updateRatingAndReview(bookId: Int, rating: Float, review: String)
+    suspend fun updateRatingAndReview(bookId: Long, rating: Float, review: String)
 
     @Query("UPDATE books SET pdfPath = :pdfPath WHERE id = :bookId")
-    suspend fun updatePdfPath(bookId: Int, pdfPath: String?)
+    suspend fun updatePdfPath(bookId: Long, pdfPath: String?)
 
     @Delete
     suspend fun deleteBook(book: BookEntity)
 
     @Query("DELETE FROM books WHERE id = :bookId")
-    suspend fun deleteBookById(bookId: Int)
+    suspend fun deleteBookById(bookId: Long)
 }
